Consolidated OcrViolationTicket model and refactored all the services in order to use the model from the common library only to fix and avoid serialization issues
