+++ src/backend/TrafficCourts/Citizen.Service/Features/Tickets/AnalyseHandler.cs
@@ -1,8 +1,8 @@
 using MediatR;
-using TrafficCourts.Citizen.Service.Models.Tickets;
 using TrafficCourts.Citizen.Service.Services;
 using TrafficCourts.Citizen.Service.Validators;
 using TrafficCourts.Common.Features.FilePersistence;
+using TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0;
 
 namespace TrafficCourts.Citizen.Service.Features.Tickets;
