+++ src/backend/TrafficCourts/Citizen.Service/Validators/Rules/CountSectionRule.cs
@@ -1,6 +1,6 @@
 using System.Text.RegularExpressions;
-using TrafficCourts.Citizen.Service.Models.Tickets;
 using TrafficCourts.Common.Features.Lookups;
+using TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0;
 
 namespace TrafficCourts.Citizen.Service.Validators.Rules;
