+++ src/backend/TrafficCourts/Citizen.Service/Validators/Rules/DriversLicenceValidRule.cs
@@ -1,5 +1,5 @@
 using System.Text.RegularExpressions;
-using TrafficCourts.Citizen.Service.Models.Tickets;
+using TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0;
 
 namespace TrafficCourts.Citizen.Service.Validators.Rules;
