+++ src/backend/TrafficCourts/Citizen.Service/Validators/Rules/FieldMatchesRegexRule.cs
@@ -1,5 +1,5 @@
 using System.Text.RegularExpressions;
-using static TrafficCourts.Citizen.Service.Models.Tickets.OcrViolationTicket;
+using TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0;
 
 namespace TrafficCourts.Citizen.Service.Validators.Rules;
 
@@ -8,7 +8,7 @@ public class FieldMatchesRegexRule : ValidationRule
     private readonly string _pattern;
     private readonly string _reason;
 
-    public FieldMatchesRegexRule(Models.Tickets.Field field, string pattern, string reason) : base(field)
+    public FieldMatchesRegexRule(Field field, string pattern, string reason) : base(field)
     {
         this._pattern = pattern;
         this._reason = reason;
