+++ src/backend/TrafficCourts/Citizen.Service/Validators/Rules/LowConfidenceGlobalRule.cs
@@ -1,5 +1,5 @@
-using TrafficCourts.Citizen.Service.Models.Tickets;
-using static TrafficCourts.Citizen.Service.Models.Tickets.OcrViolationTicket;
+using TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0;
+using static TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0.OcrViolationTicket;
 
 namespace TrafficCourts.Citizen.Service.Validators.Rules;
