+++ src/backend/TrafficCourts/Citizen.Service/Validators/Rules/TimeRule.cs
@@ -1,6 +1,4 @@
-using System;
-using TrafficCourts.Citizen.Service.Models.Tickets;
-using static TrafficCourts.Citizen.Service.Models.Tickets.OcrViolationTicket;
+using TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0;
 
 namespace TrafficCourts.Citizen.Service.Validators.Rules;
