+++ src/backend/TrafficCourts/Test/Citizen.Service/Validators/Rules/CountActRegMustBeMVATest.cs
@@ -1,10 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using TrafficCourts.Citizen.Service.Models.Tickets;
+﻿using System.Threading.Tasks;
 using TrafficCourts.Citizen.Service.Validators.Rules;
+using TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0;
 using Xunit;
 
 namespace TrafficCourts.Test.Citizen.Service.Validators.Rules;
