+++ src/backend/TrafficCourts/Test/Citizen.Service/Validators/Rules/CountSectionRuleTest.cs
@@ -5,8 +5,9 @@ using TrafficCourts.Citizen.Service.Validators;
 using TrafficCourts.Citizen.Service.Validators.Rules;
 using TrafficCourts.Common.Features.Lookups;
 using TrafficCourts.Common.Models;
+using TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0;
 using Xunit;
-using static TrafficCourts.Citizen.Service.Models.Tickets.OcrViolationTicket;
+using static TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0.OcrViolationTicket;
 
 namespace TrafficCourts.Test.Citizen.Service.Validators.Rules;
