+++ src/backend/TrafficCourts/Test/Citizen.Service/Validators/Rules/DateOfServiceLT30RuleTest.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Threading.Tasks;
-using TrafficCourts.Citizen.Service.Models.Tickets;
 using TrafficCourts.Citizen.Service.Validators.Rules;
+using TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0;
 using Xunit;
 
 namespace TrafficCourts.Test.Citizen.Service.Validators.Rules;
